java 11 should be ready to go

+++ siloCore/src/main/java/de/tum/bgu/msm/models/relocation/moves/DefaultDwellingProbabilityStrategy.java
@@ -13,6 +13,6 @@ public class DefaultDwellingProbabilityStrategy extends JavaScriptCalculator<Dou
     }
 
     public double calculateSelectDwellingProbability(double dwellingUtility) {
-        return super.calculate("calculateSelectDwellingProbability", dwellingUtility).doubleValue();
+        return super.calculate("calculateSelectDwellingProbability", dwellingUtility);
     }
 }
